Adds the map function to add the individual ids to the container underneath the input.
The key for each id added using .map is its position in the array, this is ok for now but when the the functionality for the user to delete any accidentally
added Ids this key will have to be changed as when an element in an array is removed the position of another will change thus changing the key.

The styling for the text in the id label in the group container still needs to be done. Going to work on the transition between the login screen and the create a group screen.
The invite screen still needs to be added so that the user can join already created groups instead of having to create their own group every time.